css/css3-selectors/i18n/css3-selectors-lang-002.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>:lang(es), lang="es" on parent</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#lang-pseudo'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; height: 50px; background-color: blue; }
#box:lang(es) { width: 100px; }
.ref { width: 100px; height: 50px; background-color: blue; }
</style>
</head>
<body>



<p>The following two blocks should be <strong>identical</strong>.</p>
<div class="test" lang="es"><div id="box">&#xA0;</div></div>
<p class="ref">&#xA0;</p>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "A :lang value that matches an identical lang attribute value on a parent element will produce styling.");
</script>

<div id='log'></div>

</body>
</html>
